requirements.txt
requests
pandas
numpy
biopython
matplotlib